+++ packages/react-ui/components/TokenInput/__stories__/TokenInput.stories.tsx
@@ -521,8 +521,8 @@ OnUnexpectedInputValidation.parameters = {
     skip: [
       {
         in: ['firefox', 'firefox8px', 'firefoxFlat8px', 'firefoxDark'],
-        tests: 'token select',
-        reason: 'flacky "clearedOnNullReturn"',
+        tests: ['token select', 'token edit'],
+        reason: 'flacky',
       },
     ],
     tests: {
@@ -612,12 +612,15 @@ OnUnexpectedInputValidation.parameters = {
             bridge: true,
           })
           .click(this.browser.findElement({ css: '[data-comp-name~="TokenInput"]' }))
+          .pause(300)
           .sendKeys('aaa')
+          .pause(300)
           .sendKeys(this.keys.ENTER)
-          .pause(1000)
+          .pause(300)
           .sendKeys('bbb')
+          .pause(300)
           .sendKeys(this.keys.ENTER)
-          .pause(1000)
+          .pause(300)
           .perform();
 
         await this.browser
@@ -625,10 +628,12 @@ OnUnexpectedInputValidation.parameters = {
             bridge: true,
           })
           .doubleClick(this.browser.findElement({ css: '[data-comp-name~="Token"]' }))
+          .pause(1000)
           .sendKeys('aaa')
+          .pause(300)
           .move({ x: 0, y: 0 })
           .click()
-          .pause(1000)
+          .pause(300)
           .perform();
 
         const withSameValue = await this.takeScreenshot();
@@ -638,10 +643,12 @@ OnUnexpectedInputValidation.parameters = {
             bridge: true,
           })
           .doubleClick(this.browser.findElement({ css: '[data-comp-name~="Token"]' }))
+          .pause(1000)
           .sendKeys('zzz')
+          .pause(300)
           .move({ x: 0, y: 0 })
           .click()
-          .pause(1000)
+          .pause(300)
           .perform();
 
         const withNotEditedToken = await this.takeScreenshot();
@@ -651,15 +658,17 @@ OnUnexpectedInputValidation.parameters = {
             bridge: true,
           })
           .doubleClick(this.browser.findElement({ css: '[data-comp-name~="Token"]' }))
+          .pause(1000)
           .sendKeys(this.keys.BACK_SPACE)
           .sendKeys(this.keys.BACK_SPACE)
           .sendKeys(this.keys.BACK_SPACE)
           .sendKeys(this.keys.BACK_SPACE)
           .sendKeys(this.keys.BACK_SPACE)
           .sendKeys('clear')
+          .pause(300)
           .move({ x: 0, y: 0 })
           .click()
-          .pause(1000)
+          .pause(300)
           .perform();
 
         const withRemovedToken = await this.takeScreenshot();
@@ -669,12 +678,15 @@ OnUnexpectedInputValidation.parameters = {
             bridge: true,
           })
           .doubleClick(this.browser.findElement({ css: '[data-comp-name~="Token"]' }))
+          .pause(1000)
           .sendKeys('EDITED')
+          .pause(300)
           .sendKeys(this.keys.ARROW_DOWN)
           .sendKeys(this.keys.ENTER)
+          .pause(300)
           .move({ x: 0, y: 0 })
           .click()
-          .pause(1000)
+          .pause(300)
           .perform();
 
         const withEditedToken = await this.takeScreenshot();