+++ packages/react-ui/internal/ThemePlayground/constants.ts
@@ -5,6 +5,8 @@ export enum ThemeType {
   Dark = 'dark',
   DefaultOld = 'defaultOld',
   FlatOld = 'flatOld',
+  Theme2022 = 'theme2022',
+  Theme2022Dark = 'theme2022Dark',
 }
 
 export const VARIABLES_GROUPS = [